dark theme and folder organisation
