+++ lib/gui/cross_stitch/__init__.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+# Authors: see git history
+#
+# Copyright (c) 2026 Authors
+# Licensed under the GNU GPL version 3.0 or later.  See the file LICENSE for details.
+
+from .cross_stitch_canvas import CrossStitchCanvasWindow
+
+__all__ = ["CrossStitchCanvasWindow"]
