+++ lib/gui/cross_stitch/grid_mapper.py
