+++ lib/gui/cross_stitch/grid_state.py
