+++ lib/gui/cross_stitch/region_merger.py
