+++ lib/gui/cross_stitch/undo_manager.py
